refactor(job offers): change folder name in backend
